requirements.txt
aiogram>=3.4
apscheduler
python-dotenv
aiofiles

